+++ rollup.config.js
@@ -1,10 +1,11 @@
-import { terser } from 'rollup-plugin-terser' // 代码压缩
-import typescript from '@rollup/plugin-typescript'
+import { terser } from 'rollup-plugin-terser'; // 代码压缩
+import typescript from '@rollup/plugin-typescript';
 export default {
   input: 'src/index.ts',
   output: [
     {
       file: 'lib/ja-fetch.js',
+      sourcemap: true,
     },
     {
       file: 'lib/ja-fetch.min.js',
@@ -13,4 +14,4 @@ export default {
     },
   ],
   plugins: [typescript()],
-}
+};